data Product and Categories

+++ src/main/java/de/telran/onlineshopgarden/entity/OrderItem.java
@@ -18,7 +18,7 @@ public class OrderItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long orderItemId;
+    private Integer orderItemId;
 
     @ManyToOne
     @JoinColumn(name = "order_id", nullable = false)
